Fix some warnings related to 32-bit code/emscripten

+++ include/chemfiles/parse.hpp
@@ -187,12 +187,12 @@ inline size_t scan(string_view input, Args& ...args) {
 /// of `*` characters if the integer is out of range.
 ///
 /// [hybrid36]: http://cci.lbl.gov/hybrid_36/
-std::string encode_hybrid36(uint64_t width, int64_t value);
+std::string encode_hybrid36(size_t width, int64_t value);
 
 /// Decodes an integer using the [hybrid36] encoding scheme.
 ///
 /// [hybrid36]: http://cci.lbl.gov/hybrid_36/
-int64_t decode_hybrid36(uint64_t width, string_view input);
+int64_t decode_hybrid36(size_t width, string_view input);
 
 /// Maximum value for a width 4 number
 constexpr auto MAX_HYBRID36_W4_NUMBER = 2436111;
